+++ src/main/webapp/components/basic-search/basic-search.stories.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import { storiesOf } from './@storybook/react'
+import { storiesOf } from '../../@storybook/react'
 import { action } from '@connexta/ace/@storybook/addon-actions'
 import BasicSearch from './basic-search'
