+++ src/main/webapp/components/basic-search/index.js
@@ -1,0 +1,3 @@
+import BasicSearchGQL, { BasicSearch } from './basic-search'
+
+export { BasicSearchGQL as default, BasicSearch }
