+++ src/main/webapp/components/containers/index.js
@@ -1,0 +1,3 @@
+import FilterCard from './filter-card'
+
+export { FilterCard }
