+++ src/main/webapp/components/query-schedule/index.tsx
@@ -1,0 +1,2 @@
+import QuerySchedule from './query-schedule'
+export default QuerySchedule
