+++ src/main/webapp/components/query-schedule/query-schedule.tsx
